requirements.txt
streamlit>=1.30.0

# Agent orchestration
langgraph>=0.0.40
langchain-core>=0.1.45
langchain>=0.1.0

# LLM provider (Groq)
groq>=0.9.0

# PDF parsing & generation
pypdf>=3.17.0
reportlab>=4.0.0

# Utilities
python-dotenv>=1.0.0
